fix: patch for styled-components v6.1.17 on React Native 0.79
Fixes the "ReferenceError: document is not defined" error caused by
a web-specific code snippet inside the native bundle of styled-components.

The error occurs even when using 'styled-components/native',
because the library still checks for `window` and `HTMLElement`,
which leads to access of `document`, not available in the React Native environment.

This patch forces the web environment check value to `false`.
Refs:
- https://github.com/styled-components/styled-components/issues/5576#issuecomment-2816629148